feat: bring back essential tests
